backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
langgraph>=0.1.0
langchain>=0.2.0
anthropic>=0.28.0
databricks-sql-connector>=3.0.0
sqlalchemy>=2.0.0
mlflow>=2.13.0
pandas>=2.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
dspy-ai
langgraph-checkpoint-sqlite
sqlite3
